+++ authentic.pl
@@ -1,5 +1,5 @@
 #
-# Authentic Theme 17.30 (https://github.com/qooob/authentic-theme)
+# Authentic Theme 17.31 (https://github.com/qooob/authentic-theme)
 # Copyright 2016 Ilia Rostovtsev <programming@rostovtsev.ru>
 # Licensed under MIT (https://github.com/qooob/authentic-theme/blob/master/LICENSE)
 #
@@ -165,19 +165,19 @@ sub theme_popup_prehead {
             . "\n";
         print '<link href="'
             . $gconfig{'webprefix'}
-            . '/unauthenticated/css/package.min.css?1730" rel="stylesheet" type="text/css">'
+            . '/unauthenticated/css/package.min.css?1731" rel="stylesheet" type="text/css">'
             . "\n";
         print '<script src="'
             . $gconfig{'webprefix'}
-            . '/unauthenticated/js/package.min.js?1730" type="text/javascript"></script>'
+            . '/unauthenticated/js/package.min.js?1731" type="text/javascript"></script>'
             . "\n";
         print '<script src="'
             . $gconfig{'webprefix'}
-            . '/unauthenticated/js/cgi.min.js?1730" type="text/javascript"></script>',
+            . '/unauthenticated/js/cgi.min.js?1731" type="text/javascript"></script>',
             "\n";
         print '<script src="'
             . $gconfig{'webprefix'}
-            . '/unauthenticated/js/filtermatch.min.js?1730" type="text/javascript"></script>',
+            . '/unauthenticated/js/filtermatch.min.js?1731" type="text/javascript"></script>',
             "\n";
     }
 }